+++ src/app/layout/header/components/profile-dropdown/profile-dropdown.component.html
@@ -40,9 +40,9 @@
             <a class="dropdown-item" routerLink="/user/orders">My Orders</a>
           </li> -->
           <!-- <li class="divider dropdown-divider"></li> -->
-          <li role="menuitem">
+          <!-- <li role="menuitem">
             <a class="dropdown-item" routerLink="/user/addresses">My Saved Addresses</a>
-          </li>
+          </li> -->
           <!-- <li class="divider dropdown-divider"></li>
           <li role="menuitem">
             <a class="dropdown-item" routerLink="/user/favorite-products">My Favorite Products</a>
