+++ src/app/shared/components/saved-address/saved-address.component.scss
@@ -1,0 +1,19 @@
+@import "../../scss/selected_theme_variables";
+.name {
+  font-size: 1em;
+  color: $gray-900;
+  font-weight: 700;
+}
+.mob {
+  font-size: 1em;
+  color: $gray-900;
+  font-weight: 700;
+}
+.pin{font-weight: 700;}
+
+.reviewlink {
+  cursor: pointer;
+  &:hover {
+    text-decoration: none
+  }
+}
