+++ src/app/shared/index.ts
@@ -17,6 +17,7 @@ import { ProductSliderComponent } from './components/product-slider/product-slid
 import { InnerIproductComponent } from './components/product-slider/inner-product/inner-product.component'
 // Directives
 import { ZoomableDirective } from './directives/zoomable.directive';
+import { SavedAddressComponent } from './components/saved-address/saved-address.component';
 
 @NgModule({
   declarations: [
@@ -27,6 +28,7 @@ import { ZoomableDirective } from './directives/zoomable.directive';
     ZoomableDirective,
     ProductSliderComponent,
     InnerIproductComponent,
+    SavedAddressComponent
   ],
   exports: [
     // components
@@ -41,6 +43,7 @@ import { ZoomableDirective } from './directives/zoomable.directive';
     HumanizePipe,
     ZoomableDirective,
     ProductSliderComponent,
+    SavedAddressComponent,
     NguCarouselModule
   ],
   imports: [
