--- src/app/user/components/addresses/saved-address/saved-address.component.html
@@ -1,24 +1,0 @@
-<div class="items" *ngIf="userAddresses?.length > 0">
-  <div *ngFor="let address of userAddresses">
-    <hr>
-    <div class="address-row" *ngIf="!isEditAddrPressed">
-      <div class="row">
-          <div class="col name">
-              <button class='reviewlink pl-1 mb-3' (click)="editAddress(address)">Edit</button>
-          </div>
-      </div>
-      <div class="row">
-        <div class="col name">
-          <span class="name mr-2">{{address.first_name}}</span>
-          <span class="name mr-2">{{address.last_name}}</span>
-          <span class="mob">{{address.phone}}</span>
-        </div>
-      </div>
-      <div class="address">{{address.address_line_1}} {{address.address_line_2}}
-        <span class="address-field">{{address.city}}</span>,
-        <!-- <span class="address-field"> {{address.state?.name}}</span> -->
-        <span class="address-field pin">-{{address.zip_code}}</span>
-      </div>
-    </div>
-  </div>
-</div>
