--- src/app/user/components/addresses/saved-address/saved-address.component.scss
@@ -1,12 +1,0 @@
-@import '../../../../shared/scss/selected_theme_variables';
-.name {
-  font-size: 1em;
-  color: $gray-900;
-  font-weight: 700;
-}
-.mob {
-  font-size: 1em;
-  color: $gray-900;
-  font-weight: 700;
-}
-.pin{font-weight: 700;}
