+++ src/app/user/user.component.html
@@ -12,7 +12,7 @@
       <!-- <p><a routerLink="/user/overview" routerLinkActive="active">Overview</a></p> -->
       <!-- <hr> -->
       <p>
-        <a routerLink="/user/profile" routerLinkActive="active">My Profile</a>
+        <!-- <a routerLink="/user/profile" routerLinkActive="active">My Profile</a> -->
       </p>
       <hr>
       <p>
